requirements.txt
# API & Web Requests
requests==2.31.0
google-generativeai==0.3.2
praw==7.7.1
feedparser==6.0.10
beautifulsoup4==4.12.2

# Data Processing
pandas==2.1.3
python-dotenv==1.0.0

# CLI & Utilities
click==8.1.7
colorama==0.4.6
pyperclip==1.8.2
rich==13.7.0

# Scheduling
schedule==1.2.0

# Configuration
python-decouple==3.8
pyyaml==6.0.1

# Development Tools
pytest==7.4.3
black==23.12.0
pylint==3.0.3
